pyproject.toml
[project]
name = "nonebot-plugin-rimofun"
version = "0.2.0"
description = "Use RimoChan's stuff to do some interesting things"
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-plugin-send-anything-anywhere>=0.2.7",
    "pydantic<2.0.0,>=1.10.6",
    "pypinyin>=0.48.0",
    "yinglish>=1.0.1",
    "unvcode>=1.0.0",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-rimofun"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
